backend/requirements.txt
django>=6.0,<7
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-cors-headers>=4.4
python-dotenv>=1.0
